move out error and file types
