+++ src/commands/enableDisableGitOps.ts
@@ -1,6 +1,6 @@
 import { window } from 'vscode';
 import { azureTools, isAzureProvider } from '../azure/azureTools';
-import { failed } from '../errorable';
+import { failed } from '../types/errorable';
 import { telemetry, disableConfirmations } from '../extension';
 import { fluxTools } from '../flux/fluxTools';
 import { kubernetesTools } from '../kubernetes/kubernetesTools';
