+++ src/commands/showInstalledVersions.ts
@@ -1,6 +1,6 @@
 import os from 'os';
 import { env, extensions, version, window } from 'vscode';
-import { failed } from '../errorable';
+import { failed } from '../types/errorable';
 import { GitOpsExtensionConstants } from '../extension';
 import { getAzureVersion, getFluxVersion, getGitVersion, getKubectlVersion } from '../install';
