+++ src/types/errorable.ts
