+++ src/types/fileTypes.ts
